# api/openapi/schema.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Social Theory Library API",
        "version": "3.0.7",
        "description": "Canonical catalog, editorial processes and permission-scoped reading APIs."
    },
    "paths": {
...
        "/api/catalog/admin/scholar-relations/": {
            "get": {
                "operationId": "catalog_admin_scholar_relations_list",
                "tags": [
                    "catalog"
                ],
                "security": [
                    {
                        "libraryCookie": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScholarRelationPage"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "post": {
                "operationId": "catalog_admin_scholar_relations_create",
                "tags": [
                    "catalog"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScholarRelationRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/ScholarRelationRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ScholarRelationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "libraryCookie": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScholarRelation"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DirectionEnum": {
                "enum": [
                    "directed",
                    "bidirectional",
                    "undirected"
                ],
                "type": "string",
                "description": "* `directed` - directed\n* `bidirectional` - bidirectional\n* `undirected` - undirected"
            },
...
            "RelationTypeEnum": {
                "enum": [
                    "teaching",
                    "cooperation",
                    "influence",
                    "criticism",
                    "comparative_reading",
                    "other"
                ],
                "type": "string",
                "description": "* `teaching` - teaching\n* `cooperation` - cooperation\n* `influence` - influence\n* `criticism` - criticism\n* `comparative_reading` - comparative_reading\n* `other` - other"
            },
...
            "ScholarRelation": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "source_scholar": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "target_scholar": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "source_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "target_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "source_slug": {
                        "type": "string",
                        "readOnly": true
                    },
                    "target_slug": {
                        "type": "string",
                        "readOnly": true
                    },
                    "relation_type": {
                        "$ref": "#/components/schemas/RelationTypeEnum"
                    },
                    "direction": {
                        "$ref": "#/components/schemas/DirectionEnum"
                    },
                    "summary": {
                        "type": "string"
                    },
                    "source": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "readOnly": true
                    },
                    "edit_version": {
                        "type": "string"
                    },
                    "has_unpublished_changes": {
                        "type": "boolean",
                        "readOnly": true
                    }
                },
                "required": [
                    "direction",
                    "has_unpublished_changes",
                    "id",
                    "relation_type",
                    "source",
                    "source_name",
                    "source_scholar",
                    "source_slug",
                    "status",
                    "summary",
                    "target_name",
                    "target_scholar",
                    "target_slug"
                ]
            },
            "ScholarRelationPage": {
                "type": "object",
                "properties": {
                    "count": {
                        "type": "integer"
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    },
                    "previous": {
                        "type": "string",
                        "nullable": true
                    },
                    "results": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ScholarRelation"
                        }
                    }
                },
                "required": [
                    "count",
                    "next",
                    "previous",
                    "results"
                ]
            },
            "ScholarRelationRequest": {
                "type": "object",
                "properties": {
                    "source_scholar": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "target_scholar": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "relation_type": {
                        "$ref": "#/components/schemas/RelationTypeEnum"
                    },
                    "direction": {
                        "$ref": "#/components/schemas/DirectionEnum"
                    },
                    "summary": {
                        "type": "string"
                    },
                    "source": {
                        "type": "string"
                    },
                    "edit_version": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "direction",
                    "relation_type",
                    "source",
                    "source_scholar",
                    "summary",
                    "target_scholar"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            },
            "libraryCookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "stl_access"
            }
        }
    }
}